댓글 리스트 양식, 댓글 작성 처리 기능 구현중

+++ src/main/java/com/ymw/exam/demo/controller/UsrArticleController.java
@@ -49,7 +49,7 @@ public class UsrArticleController {
 
 		ResultData<Integer> writeArticleRd = articleService.writeArticle(rq.getLoginedMemberId(), boardId, title, body);
 
-		int id = (int) writeArticleRd.getData1();
+		int id = writeArticleRd.getData1();
 
 		return Utility.jsReplace(Utility.f("%d번 글이 생성되었습니다", id), Utility.f("detail?id=%d", id));
 	}
